+++ docs/conf.py
@@ -21,7 +21,7 @@ copyright = '2020, DataKitchen'
 author = 'DataKitchen'
 
 # The full version, including alpha/beta/rc tags
-release = u'1.3.3'
+release = u'1.7.0'
 
 # -- General configuration ---------------------------------------------------